enable2FA

+++ app/repositories/user_management_repository.py
@@ -1,13 +1,13 @@
-from datetime import timedelta
 from flask import abort
+from flask_jwt_extended import create_access_token
 from flask_sqlalchemy import SQLAlchemy
 from injector import inject
-from app.dtos.register_dto import RegisterDto
+import pyotp
 from app.models.role import Role
 from app.models.user import User
 from app.models.user_token import UserToken
 from werkzeug.security import check_password_hash, generate_password_hash
-from flask_jwt_extended import create_access_token
+
 
 class UserManagementRepository:
     @inject
@@ -15,24 +15,22 @@ class UserManagementRepository:
         self.db = db
         
 
-
     def register(self, user: User):
         role = self.db.session.query(Role).filter_by(name='user').one_or_none()
         
         if not role:
             abort(404, description='The required role is not exist')
             return
-
         user.roles.append(role)
+        user.secret_key = pyotp.random_base32()
         
         self.db.session.add(user)
         self.db.session.commit()
 
 
-
     def login(self, user: User):
         user_in_db = self.get_user_by_username(user.username)
-        
+
         if not user_in_db:
             abort(401, description='Invalid credentials')
             return
@@ -41,21 +39,41 @@ class UserManagementRepository:
             abort(401, description='Invalid credentials' )
             return
 
+        if user_in_db.is_2FA_enabled:
+            return user_in_db
+                
         existing_access_token = self.db.session.query(UserToken).filter_by(user_id=user_in_db.id).one_or_none()
         if existing_access_token:
             self.db.session.delete(existing_access_token)
         
-        
         token = UserToken(
-            user_id=user_in_db.id,
+            user_id=user.id,
             name='access_token',
-            value=create_access_token(identity=user_in_db.username, expires_delta=timedelta(days=30))
+            value=create_access_token(identity=user.username)
         )
         user_in_db.tokens.append(token)
 
         self.db.session.commit()
+        
+        return
     
     
+    def login_with_otp(self, user: User):
+
+        existing_access_token = self.db.session.query(UserToken).filter_by(user_id=user.id).one_or_none()
+        if existing_access_token:
+            self.db.session.delete(existing_access_token)
+        
+        token = UserToken(
+                user_id=user.id,
+                name='access_token',
+                value=create_access_token(identity=user.username)
+            )
+        user.tokens.append(token)
+
+        self.db.session.commit()
+        
+        
     def reset_password(self, user: User, new_password):
         user.password_hash = generate_password_hash(new_password)
         self.db.session.commit() 
@@ -74,3 +92,7 @@ class UserManagementRepository:
     def verify_password(self, real_password, entered_password):
         return check_password_hash(real_password, entered_password)
     
+    
+    def enable_2FA(self, user: User):
+        user.is_2FA_enabled = True
+        self.db.session.commit()
